Add files via upload
First, the script will make a fetch request to the URL https://cfw-takehome.developers.workers.dev/api/variants, and parse the response as JSON. The response will be an array of URLs, which are saved to a variable. Then, the program will randomly choose one of the URL's within the array and save it as a cookie. The URL is then used to make a new fetch request, which will return the site's HTML code. Using the HTMLRewriter API, the HTML code is modified and personalized, then returned to Cloudflare. 
This allows the user to use the applicaion to randomly visit one of the customized url's, and experience persisting navigation so that they always go to the same url after their first visit..